Add names to differentiate views in workspace
Added the ability to add names to the shorelines table so users are able to differentiate between uploads as well as giving them the ability to remove one view in their workspace from another. The uploaded shapefile/lidar file name is also added to the view title. Had to substring the incoming parameters to get under Postgres' default 63 character table/view name limit.